Refactor on BooksRepository
- It now receives a database Interface which is responsible for touching the database whatever it is.
- Create first specs
- Move searchBook logic to the database interface (the repository is not responsible for this, for example: if you will use a SQL query on a relational database, okay. If you will parse many documents on a NoSQL database, thats all right too. The repository just expects an array of books in return)
- Creates a constructor and some instance methods on FirebaseManager (all of its functions will be instance methods soon)
